+++ src/apps/docking-viewer/viewport.tsx
@@ -54,6 +54,7 @@ function occlusionStyle(plugin: PluginContext) {
                 scale: 1.0,
                 threshold: 0.33,
                 color: Color(0x0000),
+                includeTransparent: true,
             } }
         }
     } });